@@ -12,7 +12,7 @@ import (
 
 func TestIsRewriteRequest(t *testing.T) {
 	t.Run("returns true when rewrite host exists", func(t *testing.T) {
-		ctx := context.WithValue(context.Background(), rewrite.RewriteHostKey, "example.com")
+		ctx := context.WithValue(t.Context(), rewrite.RewriteHostKey, "example.com")
 		request := &contracts.Request{}
 
 		result := rewrite.IsRewriteRequest(
@@ -26,7 +26,7 @@ func TestIsRewriteRequest(t *testing.T) {
 		request := &contracts.Request{}
 
 		result := rewrite.IsRewriteRequest(
-			request.WithContext(context.Background()),
+			request.WithContext(t.Context()),
 		)
 
 		assert.False(t, result)
@@ -36,7 +36,7 @@ func TestIsRewriteRequest(t *testing.T) {
 func TestGetRewriteHost(t *testing.T) {
 	t.Run("returns host when exists", func(t *testing.T) {
 		expected := "example.com"
-		ctx := context.WithValue(context.Background(), rewrite.RewriteHostKey, expected)
+		ctx := context.WithValue(t.Context(), rewrite.RewriteHostKey, expected)
 		request := &contracts.Request{}
 
 		result, err := rewrite.GetRewriteHost(
@@ -51,7 +51,7 @@ func TestGetRewriteHost(t *testing.T) {
 		request := &contracts.Request{}
 
 		result, err := rewrite.GetRewriteHost(
-			request.WithContext(context.Background()),
+			request.WithContext(t.Context()),
 		)
 
 		require.NoError(t, err)
@@ -59,7 +59,7 @@ func TestGetRewriteHost(t *testing.T) {
 	})
 
 	t.Run("returns error when host has invalid type", func(t *testing.T) {
-		ctx := context.WithValue(context.Background(), rewrite.RewriteHostKey, 123)
+		ctx := context.WithValue(t.Context(), rewrite.RewriteHostKey, 123)
 		request := &contracts.Request{}
 
 		result, err := rewrite.GetRewriteHost(
