@@ -3,14 +3,12 @@ package main
 import (
 	"os"
 
-	"github.com/evg4b/uncors/internal/tui"
-
-	"github.com/evg4b/uncors/internal/config/validators"
-
 	"github.com/charmbracelet/log"
 	"github.com/evg4b/uncors/internal/config"
+	"github.com/evg4b/uncors/internal/config/validators"
 	"github.com/evg4b/uncors/internal/helpers"
 	"github.com/evg4b/uncors/internal/infra"
+	"github.com/evg4b/uncors/internal/tui"
 	"github.com/evg4b/uncors/internal/uncors"
 	"github.com/evg4b/uncors/internal/version"
 	"github.com/fsnotify/fsnotify"
